requirements.txt
apex==0.1
cycler==0.10.0
dataclasses==0.6
decorator==4.4.2
ffmpeg==1.4
future==0.18.2
imageio==2.9.0
kiwisolver==1.3.1
matplotlib==3.3.3
natsort==7.1.0
networkx==2.5
numpy==1.19.4
opencv-python==4.4.0.46
Pillow==8.0.1
protobuf==3.14.0
pyparsing==2.4.7
python-dateutil==2.8.1
PyWavelets==1.1.1
scikit-image==0.17.2
scipy==1.5.4
six==1.15.0
tensorboardX==2.1
tifffile==2020.11.18
torch==1.7.0
torchvision==0.8.1
tqdm==4.53.0
typing-extensions==3.7.4.3
